Start migration to a simpler SettingsProvider
