--- lib/controller/cubit/theme_state.dart
@@ -1,10 +1,0 @@
-part of 'theme_cubit.dart';
-
-class ThemeState extends Equatable {
-  const ThemeState(this.themeMode);
-
-  final ThemeMode themeMode;
-
-  @override
-  List<Object> get props => [themeMode];
-}
